backend/requirements.txt
aiohttp==3.8.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.2
gunicorn==20.1.0
joblib==1.2.0
marshmallow==3.18.0
mysql-connector==2.2.9
PyMySQL==1.0.2
pytest==7.2.0
